Fix doctest and renaming src
