+++ clippy_lints/src/methods/manual_saturating_arithmetic.rs
